Trpc - refactored logic into separate routers, updated text shown in previous chats button
