+++ package.json
@@ -32,10 +32,10 @@
 		"@t3-oss/env-nextjs": "^0.10.1",
 		"@tanstack/react-query": "^5.40.0",
 		"@tanstack/react-query-devtools": "^5.40.0",
-		"@trpc/client": "11.0.0-rc.377",
-		"@trpc/next": "11.0.0-rc.377",
-		"@trpc/react-query": "11.0.0-rc.377",
-		"@trpc/server": "11.0.0-rc.377",
+		"@trpc/client": "11.0.0-rc.390",
+		"@trpc/next": "11.0.0-rc.390",
+		"@trpc/react-query": "11.0.0-rc.390",
+		"@trpc/server": "11.0.0-rc.390",
 		"@types/uuid": "^9.0.8",
 		"@vercel/analytics": "^1.2.2",
 		"@vercel/speed-insights": "^1.0.10",
