+++ src/app/_trpc/client.ts
@@ -1,5 +1,5 @@
 import { createTRPCReact } from '@trpc/react-query'
 
-import { AppRouter } from '@/trpc'
+import { AppRouter } from '@/server/trpc/routers'
 
 export const trpc = createTRPCReact<AppRouter>({})
