+++ src/app/api/trpc/[trpc]/route.ts
@@ -1,6 +1,6 @@
 import { fetchRequestHandler } from '@trpc/server/adapters/fetch'
 
-import { appRouter } from '@/trpc'
+import { appRouter } from '@/server/trpc/routers'
 
 // This function can run for a maximum of 5 seconds
 export const maxDuration = 60
